+++ Qtica/utils/color.py
@@ -1,11 +1,60 @@
+#!/usr/bin/python3
 # coding:utf-8
+
 from PySide6.QtGui import QColor
+from collections import Counter
 from random import random
+from PIL import Image
 from re import split
+from typing import Any
 
 
-# def imageAverageColor(image: str):
-#     ...
+class DetectImageColors:
+    def __init__(self, image: str):
+
+        self._image_path = image
+        self._image = Image.open(image)
+
+        # Resize the image to improve performance
+        self._image = self._image.resize((100, 100))
+        self._image = self._image.convert('RGB')
+
+        # Count the occurrences of each color
+        self._color_counter = Counter(self._image.getdata())
+
+    @property
+    def image_path(self) -> str:
+        return self._image_path
+
+    @property
+    def total(self):
+        return self._color_counter.total()
+
+    @property
+    def count(self):
+        return self._color_counter.__len__()
+
+    @property
+    def colors(self) -> Any:
+        return self._image.getdata()
+
+    @property
+    def most_common(self, count: int = None) -> list[tuple[tuple[int], int]]:
+        '''
+        count: int
+            count == None, default = cls.total
+        '''
+        return self._color_counter.most_common(count
+                                               if count is not None
+                                               else self._color_counter.total())
+
+
+def get_image_average_color(image: str, 
+                            size: tuple = (100, 100)) -> list[int]:
+    img = Image.open(image)
+    img = img.convert("RGB")
+    img = img.resize(size)
+    return img.getpixel(tuple(map(lambda x: x  // 2, size)))
 
 
 def mixColor(c1: QColor, c2: QColor, weight: float) -> QColor:
@@ -56,21 +105,25 @@ def translucent(color: QColor, alpha: int) -> QColor:
     return QColor(color.red(), color.green(), color.blue(), alpha)
 
 
-def get_color_from_hex(s):
+def get_color_from_hex(color: str):
     '''Transform a hex string color to a kivy
     :class:`~kivy.graphics.Color`.
     '''
-    if s.startswith('#'):
-        return get_color_from_hex(s[1:])
+
+    if color.startswith('#'):
+        return get_color_from_hex(color[1:])
 
     value = [int(x, 16) / 255.
-             for x in split('([0-9a-f]{2})', s.lower()) if x != '']
+             for x in split('([0-9a-f]{2})', color.lower()) 
+             if x != '']
+
     if len(value) == 3:
         value.append(1.0)
+
     return value
 
 
-def get_hex_from_color(*rgb: list[int]):
+def get_hex_from_color(*rgb: list[int]) -> str:
     '''Transform a rgb(0, 0, 0) color to hex value::
         >>> get_hex_from_color((0, 1, 0))
         '#00ff00'
@@ -81,22 +134,36 @@ def get_hex_from_color(*rgb: list[int]):
     return '#' + ''.join(['{0:02x}'.format(int(x)) for x in rgb])
 
 
-def get_random_color(alpha=1.0):
+def get_random_color(alpha: float = 1.0) -> list[int]:
     '''Returns a random color (4 tuple).
 
     :Parameters:
         `alpha`: float, defaults to 1.0
             If alpha == 'random', a random alpha value is generated.
     '''
-    return [random(), random(), random(), random() if alpha == 'random' else alpha]
+    return [random() * 255,
+            random() * 255, 
+            random() * 255, 
+            random() * 255 if alpha == 'random' else alpha]
 
 
-def is_color_transparent(c):
+def is_color_transparent(*rgb: list[int]):
     '''Return True if the alpha channel is 0.'''
-    if len(c) < 4:
+    if len(rgb) < 4:
         return False
 
-    if float(c[3]) == 0.:
+    if float(rgb[3]) == 0.:
         return True
 
     return False
+
+
+def print_hex_color(*rgb):
+    '''
+    arg: list[int]
+        - (int, int, int, int)
+        - (int, int, int)
+
+    Return colored text.
+    '''
+    print('\033[38;2;{};{};{}m{}\033[0m'.format(rgb[0], rgb[1], rgb[2], rgb))