+++ Qtica/widgets/line_edit.py
@@ -1,12 +1,11 @@
 from PySide6.QtCore import Qt
-from PySide6.QtGui import QAction
 from PySide6.QtWidgets import QLineEdit
 from ..enums.events import EventTypeVar
 from ..enums.signals import SignalTypeVar
 from ..core.base import WidgetBase
 
 
-class Label(WidgetBase, QLineEdit):
+class LineEdit(WidgetBase, QLineEdit):
     def __init__(self, 
                  uid: str = None, 
                  signals: SignalTypeVar = None, 
@@ -17,5 +16,3 @@ class Label(WidgetBase, QLineEdit):
                  **kwargs):
         QLineEdit.__init__(self)
         super().__init__(uid, signals, events, qss, attrs, flags, **kwargs)
-
-        # self.addAction(QLineEdit.ActionPosition)