+++ examples/get_start.py
@@ -1,22 +1,3 @@
-<p align="center">
-  <a href="https://omamkaz.gitbook.io/qtica/">
-    <img alt="Qtica" src="./logo.png">
-  </a>
-</p>
-
-<p align="center">
-  A Fast Way to Done Your Idea!
-</p>
-
-Qtica is a Python library that provides a lightweight API around native PySide6, allowing for lightning-fast GUI prototyping using modern declarative UI techniques entirely within Python.
-
-# Documention
-
-[Learn More About](https://omamkaz.gitbook.io/qtica/)
-
-# Qtica Get Start
-
-```python
 #!/usr/bin/python3
 
 import os
@@ -32,11 +13,11 @@ from Qtica.widgets import (
   Application
 )
 
-
 class Window(BehaviorDeclarative):
   def update_background(self):
     bg_color = get_hex_from_color(*get_random_color())
     fg_color = get_hex_from_color(*get_random_color())
+
     Api.fetch("window").qss.update({"background-color": bg_color})
     Api.fetch("label").qss.update({"color": fg_color})
 
@@ -60,4 +41,3 @@ app = Application(sys.argv)
 window = Window()
 window.show()
 sys.exit(app.exec())
-```